generate-2fa/requirements.txt
pyotp
qrcode
psycopg2-binary
cryptography
Pillow
qrcode[pil]
